Extend Base.setindex
Import Base.setindex, so that this function is extended, instead of
creating a new funciton SIMD.setindex.

This makes it possible to use SIMD.jl together with e.g. StaticArrays.jl
without having to qualify the setindex function on each use.

This commit also fixes a bug in the argument order when the index is
not an Int.